40 pin 16 bit mode working on A500

+++ sdrr/include/test/stub.h
@@ -31,17 +31,17 @@ extern void stub_log(const char* msg, ...);
 
 // Externs required by firmware
 #define RAM_ROM_TABLE_SIZE (1024 * 1024)
-extern uint32_t *get_ram_rom_image_table_aligned();
+extern uint64_t *get_ram_rom_image_table_aligned();
 #if defined(TEST_MAIN_C)
 // Allocate twice the required RAM ROM table size, so it can be aligned to
 // 512KB (done in preload_rom_image).
 uint32_t test_ram_rom_image_table[RAM_ROM_TABLE_SIZE*2/4] = {0};
-uint32_t *get_ram_rom_image_table_aligned() {
+uint64_t *get_ram_rom_image_table_aligned() {
     uint64_t address = (uint64_t)(uintptr_t)test_ram_rom_image_table;
     address += RAM_ROM_TABLE_SIZE-1;
     address /= RAM_ROM_TABLE_SIZE;
     address = address * RAM_ROM_TABLE_SIZE;
-    return (uint32_t *)(uintptr_t)address;
+    return (uint64_t *)(uintptr_t)address;
 }
 #endif
 #define _metadata_start onerom_metadata_header
